requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
mssql-django==1.4
python-decouple==3.8
django-extensions==3.2.3
django-filter==23.5
pyodbc==5.0.1
django-import-export==4.3.9
tablib==3.8.0
openpyxl==3.1.5
